David laptop

+++ AboutNow/Controllers/JournalsController.cs
@@ -30,7 +30,7 @@ namespace AboutNow.Controllers
         // HttpGet implicit
         //Pentru fiecare jurnal se afiseaza si utilizatorul care a postat jurnalul
 
-        [Authorize(Roles = "User,Editor,Admin")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Index()
         {
             var journals = db.Journals.Include("Category").Include("User");
@@ -49,16 +49,30 @@ namespace AboutNow.Controllers
         // Se afiseaza un singur articol in functie de id-ul sau impreuna cu categoria din
         // care face parte 
         // si o sa aiba HTTPGet implicit
-        [Authorize(Roles = "User,Editor,Admin")]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Show(int id)
         {
             Journal journal = db.Journals.Include("Category").Include("Comments").Include("User")
                                                   .Where(jrn => jrn.Id == id)
                                                   .First();
 
-
+            SetAccessRights();
 
             return View(journal);
+        }
+
+        private void SetAccessRights()
+        {
+            ViewBag.AfisareButoane = false;
+
+            if (User.IsInRole("User"))
+            {
+                ViewBag.AfisareButoane = true;
+            }
+
+            ViewBag.UserCurent = _userManager.GetUserId(User);
+
+            ViewBag.EsteAdmin = User.IsInRole("Admin");
         }
 
         [HttpPost]
@@ -89,6 +103,9 @@ namespace AboutNow.Controllers
 
         // Se afiseaza formularul in care se vor completa datele unui jurnalul
         // impreuna cu selectarea categoriei din care face parte jurnalul
+        // Doar utilizatorii cu rolul editor sau admin pot adauga jurnale in platforma
+
+        [Authorize(Roles = "User,Admin")]
         public IActionResult New()
         {
             Journal journal = new Journal();
@@ -101,11 +118,12 @@ namespace AboutNow.Controllers
 
         //adaugarea jurnalului in baza de date
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult New(Journal journal)
 
         {
             journal.Date = DateTime.Now;
-            journal.Categ = GetAllCategories();
+            journal.UserId = _userManager.GetUserId(User);
 
             if (ModelState.IsValid)
             {
@@ -116,11 +134,12 @@ namespace AboutNow.Controllers
             }
             else
             {
+                journal.Categ = GetAllCategories();
                 return View(journal);
             }
         }
 
-
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Edit(int id)
         {
             Journal journal = db.Journals.Include("Category")
@@ -130,46 +149,79 @@ namespace AboutNow.Controllers
             journal.Categ = GetAllCategories();
 
 
+            if (journal.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(journal);
+            }
 
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                return RedirectToAction("Index"); 
+            }
 
-            return View(journal);
+            
 
         }
 
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
         public IActionResult Edit(int id, Journal requestJournal)
         {   // in requestJournal am valorile modificate, care se salveaza
             // in variabila asta pt ca le am in edit si de acolo le modific.
 
             Journal journal = db.Journals.Find(id);
-            requestJournal.Categ = GetAllCategories();
+            
 
             if (ModelState.IsValid)
             {
-                journal.Title = requestJournal.Title;
-                journal.Content = requestJournal.Content;
-                journal.Date = requestJournal.Date;
-                journal.CategoryId = requestJournal.CategoryId;
-                db.SaveChanges();
-                TempData["message"] = "Jurnal Editat cu Succes";
-                return RedirectToAction("Index");
+                if (journal.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+                {
+                    journal.Title = requestJournal.Title;
+                    journal.Content = requestJournal.Content;
+                    journal.Date = requestJournal.Date;
+                    journal.CategoryId = requestJournal.CategoryId;
+                    db.SaveChanges();
+                    TempData["message"] = "Jurnal Editat cu Succes";
+                    return RedirectToAction("Index");
+                }
+
+
+                else
+                {
+                    TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui articol care nu va apartine";
+                    return RedirectToAction("Index");
+                }
+
             }
             else
             {
+                requestJournal.Categ = GetAllCategories();
                 return View(requestJournal);
             }
         }
 
 
         [HttpPost]
+        [Authorize(Roles = "User,Admin")]
 
         public IActionResult Delete(int id)
         {
-            Journal journal = db.Journals.Find(id);
-            db.Journals.Remove(journal);
-            db.SaveChanges();
-            TempData["message"] = "Jurnalul a fost sters cu succes";
-            return RedirectToAction("Index");
+            Journal journal = db.Journals.Include("Comments")
+                                        .Where(jrn => jrn.Id == id)
+                                        .First();
+            if (journal.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                db.Journals.Remove(journal);
+                db.SaveChanges();
+                TempData["message"] = "Jurnalul a fost sters cu succes";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari asupra unui jurnal care nu va apartine";
+                return RedirectToAction("Index");
+            }
         }
 
         [NonAction]
